requirements.txt
albumentations==1.2.1
alive-progress==2.4.0
batchgenerators==0.24
einops==0.6.0
grad-cam==1.4.6
imageio==2.32.0
loguru==0.6.0
matplotlib==3.5.2
mmcv-full==1.7.0
monai==1.0.1
networkx==3.1
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
ptflops==0.6.9
tabulate==0.8.10
thop==0.0.31.post2005241907
timm==0.6.12
torchinfo
warmup_scheduler